requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.1.0
Werkzeug==2.3.7
matplotlib==3.7.2
seaborn==0.12.2
reportlab==4.0.4
Pillow==10.0.0
python-dotenv==1.0.0
email-validator==2.0.0 